# Conflicts:
#	composer.json
#	composer.lock
#	config/app.php

+++ app/Http/Controllers/Api/Mailbox/MailgunDomainBounceController.php
@@ -4,16 +4,21 @@ namespace App\Http\Controllers\Api\Mailbox;
 
 use App\Eco\Mailbox\MailgunDomain;
 use Carbon\Carbon;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Mailgun\Mailgun;
 
 class MailgunDomainBounceController
 {
+    use AuthorizesRequests;
+
     public function index(MailgunDomain $mailgunDomain)
     {
+        $this->authorize('update', MailgunDomain::class);
+
         $mailgunClient = Mailgun::create($mailgunDomain->secret, 'https://' . config('services.mailgun.endpoint'));
 
-        $fetchedBounces = $mailgunClient->suppressions()->bounces()->index('mg.kolenberg.net')->getItems();
+        $fetchedBounces = $mailgunClient->suppressions()->bounces()->index($mailgunDomain->domain)->getItems();
 
         return collect($fetchedBounces)->map(function ($bounce) {
             return [
@@ -26,6 +31,8 @@ class MailgunDomainBounceController
 
     public function create(MailgunDomain $mailgunDomain, Request $request)
     {
+        $this->authorize('update', MailgunDomain::class);
+
         $mailgunClient = Mailgun::create($mailgunDomain->secret, 'https://' . config('services.mailgun.endpoint'));
 
         $mailgunClient->suppressions()->bounces()->create($mailgunDomain->domain, $request->input('address'), [
@@ -35,6 +42,8 @@ class MailgunDomainBounceController
 
     public function delete(MailgunDomain $mailgunDomain, $address)
     {
+        $this->authorize('update', MailgunDomain::class);
+
         $mailgunClient = Mailgun::create($mailgunDomain->secret, 'https://' . config('services.mailgun.endpoint'));
 
         $mailgunClient->suppressions()->bounces()->delete($mailgunDomain->domain, $address);
